Update: Improved web app user experience
