--- src/templates/index2.html
@@ -1,122 +1,0 @@
-<!DOCTYPE html>
-<html>
-
-	<head>
-		<title>Dexter Controller</title>
-		<meta name="viewport" content="width=device-width, initial-scale=1">
-	</head>
-
-<style>
-table {width:100%; margin-left:auto; margin-right:auto; max-width:500px; max-height:40%}
-img {width:80%; text-align:center;}
-input {width:80%;}
-td {text-align:center; vertical-align:middle;}
-</style>
-
-<script src="//ajax.googleapis.com/ajax/libs/jquery/1.9.1/jquery.min.js"></script>
-<script type=text/javascript>
-        $(function() {
-          $('a#ccw').bind('click', function() {
-            $.getJSON('/ccw',
-                function(data) {
-              //do nothing
-            });
-            return false;
-          });
-          $('a#forward').bind('click', function() {
-            $.getJSON('/forward',
-                function(data) {
-              //do nothing
-            });
-            return false;
-          });
-        });
-</script>
-	</body>
-	<div class="video_stream">
-		<img src="{{ url_for('video_feed') }}" style="width:100%; height:100%; border-radius:10px; max-width:500px; text-align:center; "></img>
-	</div>
-	<div class="remote_control">
-		<table border="0">
-			<tr>
-				<td>
-					<form action="/7" method="POST">
-							<a href=# id=ccw>
-							<img src="../static/img/ccw.svg">
-							</a>
-							</br>
-					</form>
-				</td>
-				<td>
-					<form action="/2" method="POST">
-						<a href=# id=forward>
-							<img src="../static/img/forward.svg">
-						</a>
-						</br>
-					</form>
-				</td>
-				<td>
-					<form action="/6" method="POST">
-						<a href=# id=cw>
-							<img src="../static/img/cw.svg">
-						</a>
-						</br>
-					</form>
-				</td>
-			</tr>
-			<tr>
-				<td>
-					<form action="/1" method="POST">
-						<a href=# id=left>
-							<img src="../static/img/left.svg">
-						</a>
-						</br>
-					</form>
-				</td>
-				<td>
-					<form action="/5" method="POST">
-						<a href=# id=stop>
-							<img src="../static/img/stop_2.svg">
-						</a>
-						</br>
-					</form>
-				</td>
-				<td>
-					<form action="/3" method="POST">
-						<a href=# id=right>
-							<img src="../static/img/right.svg">
-						</a>
-						</br>
-					</form>
-				</td>
-			</tr>
-			<tr>
-				<td>
-					<form action="/8" method="POST">
-						<a href=# id=speed_down>
-							<img src="../static/img/minus.svg">
-						</a>
-						</br>
-					</form>
-				</td>
-				<td>
-					<form action="/4" method="POST">
-						<a href=# id=backward>
-							<img src= "../static/img/backward.svg">
-						</a>
-						</br>
-					</form>
-				</td>
-				<td>
-					<form action="/9" method="POST">
-						<a href=# id=speed_up>
-							<img src="../static/img/plus.svg">
-						</a>
-						</br>
-					</form>
-				</td>
-			</tr>
-		</table>
-	</div>
-	</body>
-</html>
